// types.ts
export interface CardData {
  institutionName: string;
  institutionAddress: string;
  institutionWebsite: string;
  studentName: string;
  studentId: string;
  dob: string;
  course: string;
  address: string;
  mobile: string;
  email: string;
  issueDate: string;
  expiryDate: string;
  photoUrl: string;
  logoUrl: string;
  signatureUrl: string;
}